feat: adds API docs, correct error handling on RuleTracker and refactors rules into its own module
We add a draft of the API docs and refactored `Rule` into its own module
this way we reduce the public API surface area and improve readability.

Furthemore, we do some checks before inserting in `RuleTracker` making
sure that in case of rule exhaustion or other non-predictable errors we
don't leave `RuleTracker` in a partial state.